Removed symlinks (tests were not working)
